% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string-concatenation.R
\name{chr_collapse}
\alias{chr_collapse}
\title{Collapse a character vector to a single string}
\usage{
chr_collapse(strings, separator = "")
}
\arguments{
\item{strings}{A character vector, where each element of the vector is a character string.}

\item{separator}{A character string to separate the concatenated elements. Defaults to an empty string,
which results in no separation between elements.}
}
\value{
A single character string
}
\description{
Collapse a character vector to a single string
}
\examples{
chr_collapse(c("Hello", "world"), " ")
#> [1] "Hello world"
chr_collapse(letters, "-")
#> [1] "a-b-c-d-e-f-g-h-i-j-k-l-m-n-o-p-q-r-s-t-u-v-w-x-y-z"
chr_collapse(1:9)
#> [1] "123456789"
}
\seealso{
\code{\link[=str_concat]{str_concat()}} for concatenating multiple character strings/vectors together.

\code{\link[=paste0]{paste0()}} which \code{chr_collapse()} wraps around.
}
