% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector-pattern-matching.R
\name{chr_count_matches}
\alias{chr_count_matches}
\title{Count matching indices in character vector}
\usage{
chr_count_matches(strings, pattern, fixed = FALSE)
}
\arguments{
\item{strings}{A character vector, where each element of the vector is a character string.}

\item{pattern}{A single character string to be searched for in each element of \code{strings}.
By default, \code{pattern} is interpreted as a regular expression (regex). If the \code{fixed} argument is set to \code{TRUE},
\code{pattern} will be treated as a literal string to be matched exactly.}
}
\value{
A single integer representing the number of matches found
}
\description{
Count matching indices in character vector
}
\examples{
chr_count_matches(letters, "[aeiou]")
#> [1] 5
}
