% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector-organisation.R
\name{chr_unique}
\alias{chr_unique}
\title{Remove duplicated strings}
\usage{
chr_unique(strings)
}
\arguments{
\item{strings}{A character vector, where each element of the vector is a character string.}
}
\value{
A character vector, usually shorter than \code{strings},
containing only distinct values.
}
\description{
Remove duplicate values from a character vector.
}
\examples{
chr_unique(c("a", "b", "a", "c"))
#> [1] "a" "b" "c"
}
\seealso{
\code{\link[=unique]{unique()}} which this function wraps around
}
