% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector-pattern-matching.R
\name{chr_which}
\alias{chr_which}
\title{Find matching indices in character vector}
\usage{
chr_which(string, pattern)
}
\arguments{
\item{string}{A character vector of strings to search in.}

\item{pattern}{The pattern to look for.}
}
\value{
An integer vector of matching indices in string.
}
\description{
Find matching indices in character vector
}
\examples{
chr_which(letters, "[aeiou]")
#> [1]  1  5  9 15 21
}
