% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string-concatenation.R
\name{str_concat}
\alias{str_concat}
\title{Concatenate strings}
\usage{
str_concat(..., separator = "")
}
\arguments{
\item{...}{One or more character vectors or objects coercible to character vectors. These are the strings
or objects to be concatenated.}

\item{separator}{A character string to separate the concatenated elements. Defaults to an empty string,
which results in no separation between elements.}
}
\value{
A single character string representing the concatenation of all input elements, separated by
the specified \code{separator}.
}
\description{
\code{str_concat()} takes multiple string arguments and concatenates them into a single string,
inserting a specified separator between each piece.
}
\examples{
str_concat("Hello", "world", separator = " ")
#> [1] "Hello world"
str_concat("2023", "01", "01", separator = "-")
#> [1] "2023-01-01"
str_concat("One", "Two", "Three")
#> [1] "OneTwoThree"

}
\seealso{
\code{\link[=chr_collapse]{chr_collapse()}} for reduce a character vector into a single string.

\code{\link[=paste]{paste()}}, which \code{str_concat()} wraps around.
}
