% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector-pattern-matching.R
\name{chr_count_matches}
\alias{chr_count_matches}
\title{Count matching indices}
\usage{
chr_count_matches(string, pattern)
}
\arguments{
\item{string}{A character vector of strings to search in.}

\item{pattern}{The pattern to look for.}
}
\value{
A single integer representing the number of matches found
}
\description{
Count matching indices
}
\examples{
chr_count_matches(letters, "[aeiou]")
#> [1] 5
}
