% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string-transformation.R
\name{str_indent}
\alias{str_indent}
\title{Add characters to the ends of strings}
\usage{
str_indent(
  strings,
  indent_width,
  indent = " ",
  side = c("left", "right", "both")
)
}
\arguments{
\item{strings}{A character vector, where each element of the vector is a character string.}

\item{indent_width}{The number of times the \code{indent} should be repeated, on each side specified}

\item{indent}{The characters to be used to indent.}

\item{side}{Which side of the string should the indent be applied, one of
"left", "right" or "both". The left side is the default.}
}
\value{
A vector similar to \code{strings}, but with the
}
\description{
Add characters to the ends of strings
}
\examples{
str_indent(c("Hello", "World"), 3)
#> [1] "   Hello" "   World"
str_indent(c("Hello", "World"), 3, ".")
#> [1] "...Hello" "...World"
str_indent(c("Hello", "World"), 3, "-", "both")
#> [1] "---Hello---" "---World---"
}
