% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string-transformation.R
\name{str_length}
\alias{str_length}
\alias{str_width}
\title{Get the length or width of strings}
\usage{
str_length(strings)

str_width(strings)
}
\arguments{
\item{strings}{A character vector, where each element of the vector is a character string.}
}
\value{
A numeric vector the same length as string.

\code{str_length()}: The number of characters in a string.

\code{str_width()}: The amount of space the string will take up when printed with
a fixed-width font, such as in the console.
}
\description{
Get the length in characters or width of strings, which differ where some
characters take up multiple characters worth of space, like emojis.
}
\examples{
str_length(c("Hello", "everyone"))
#> [1] 5 8

str_length("😊")
#> [1] 1
str_width("😊")
#> [1] 2
}
\seealso{
Base \code{\link[=nchar]{nchar()}} which these functions are built with.
}
