% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string-transformation.R
\name{str_to_upper_case}
\alias{str_to_upper_case}
\alias{str_to_lower_case}
\alias{str_to_snake_case}
\title{Convert strings to different cases}
\usage{
str_to_upper_case(strings)

str_to_lower_case(strings)

str_to_snake_case(strings, split_on_capital = FALSE)
}
\arguments{
\item{strings}{A character vector, where each element of the vector is a character string.}

\item{split_on_capital}{\code{str_to_snake_case()} only: whether to treat every upper case letter as
the start of a new word. This is for converting from PascalCase and
camelCase.}
}
\value{
\code{str_to_upper_case()}:
Returns \code{strings} with all lower case letters converted to upper case.

\code{str_to_lower_case()}
Returns \code{strings} with all upper case letters converted to lower case.

\code{str_to_snake_case()}
Returns \code{strings} with all upper case letters converted to lower case,
non-alphanumeric characters between words converted to an underscore "_",
and all preceding/trailing non-alphanumeric characters removed.
}
\description{
Convert strings to "UPPER CASE", "lower case" or "snake_case".
Snake case is particularly useful for clean and consistent \code{names()}.
These functions are an example of assignment in snake case.
}
\examples{
string <- "A good day for kite_flying!"

str_to_upper_case(string)
#> [1] "A GOOD DAY FOR KITE-FLYING!"
str_to_lower_case(string)
#> [1] "a good day for kite-flying!"
str_to_snake_case(string)
#> [1] "a_good_day_for_kite_flying"

# Optionally convert from PascalCase and camelCase:
str_to_snake_case(
  "AGoodDayForKiteFlying",
  split_on_capital = TRUE
)
#> [1] "a_good_day_for_kite_flying"
}
\seealso{
\code{\link[=str_replace_all]{str_replace_all()}} for replacing specific characters with other characters.
}
