% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string-transformation.R
\name{str_trim}
\alias{str_trim}
\alias{str_squish}
\title{Remove trailing and leading whitespace}
\usage{
str_trim(strings)

str_squish(strings)
}
\arguments{
\item{strings}{A character vector, where each element of the vector is a character string.}
}
\value{
The same character vector, but with whitespace removed from the start and end
of each string.
}
\description{
Remove trailing and leading whitespace
}
\examples{
print(str_trim("    Remove these extra spaces    "))
#> [1] "Remove these extra spaces"
print(str_trim("Remove trailing new line character \n"))
#> [1] "Remove trailing new line character"
print(str_trim("But leave space     in   the      middle"))
#> [1] "But leave space     in   the      middle"

print(str_squish(" Don't leave   space    in  the   middle   here  "))
#> [1] "Don't leave space in the middle here"
}
