% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string-transformation.R
\name{str_truncate}
\alias{str_truncate}
\title{Shorten a string to a specified size}
\usage{
str_truncate(
  strings,
  length,
  side = c("right", "left", "center"),
  ellipsis = "..."
)
}
\arguments{
\item{strings}{A character vector, where each element of the vector is a character string.}

\item{length}{The maximum number of characters per string in the output vector.}

\item{side}{Where the truncation should happen i.e. the location to remove characters from.}

\item{ellipsis}{Characters to replace removed characters, indicating shortening, an ellipsis
"..." by default}
}
\description{
Shorten a string to a specified size
}
\examples{
str_truncate("Help me make this shorter", 15)
#> [1] "Help me make..."
str_truncate("Help me make this shorter", 15, "left")
#> [1] "...this shorter"
str_truncate("Help me make this shorter", 15, "center")
#> [1] "Help m...horter"
}
