% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern_matching.R
\name{str_which}
\alias{str_which}
\title{Find matching indices}
\usage{
str_which(string, pattern)
}
\arguments{
\item{string}{A character vector of strings to search in.}

\item{pattern}{The pattern to look for.}
}
\value{
An integer vector of matching indices in string.
}
\description{
Find matching indices
}
\examples{
str_which(letters, "[aeiou]")
#> [1]  1  5  9 15 21
}
