% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector-pattern-matching.R
\name{chr_subset}
\alias{chr_subset}
\title{Get matching elements}
\usage{
chr_subset(strings, pattern, fixed = FALSE)
}
\arguments{
\item{strings}{A character vector from which elements are to be subset.}

\item{pattern}{A character string containing a regular expression (or a fixed string) to match against each element of \code{strings}.}
}
\value{
A character vector containing only the elements of \code{strings} that match the specified \code{pattern}. If no matches are found, an empty character vector is returned.
}
\description{
\code{chr_subset} returns a subset of the input character vector, containing only the elements that match a specified pattern.
}
\examples{
chr_subset(c("apple", "banana", "cherry", "date"), "a")
#> [1] "apple" "banana" "date"
chr_subset(c("apple", "banana", "cherry", "date"), "^a")
#> [1] "apple"

}
