% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector-pattern-matching.R
\name{chr_which}
\alias{chr_which}
\title{Find matching indices in character vector}
\usage{
chr_which(strings, pattern, fixed = FALSE)
}
\arguments{
\item{strings}{A character vector, where each element of the vector is a character string.}

\item{pattern}{A single character string to be searched for in each element of \code{strings}.
By default, \code{pattern} is interpreted as a regular expression (regex). If the \code{fixed} argument is set to \code{TRUE},
\code{pattern} will be treated as a literal string to be matched exactly.}

\item{fixed}{Logical; whether \code{pattern} should be matched exactly,
treating regex special characters as regular  string characters. Default \code{FALSE}.}
}
\value{
An integer vector of matching indices in string.
}
\description{
Find matching indices in character vector
}
\examples{
chr_which(letters, "[aeiou]")
#> [1]  1  5  9 15 21
}
