% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string-pattern-matching.R
\name{str_detect_match}
\alias{str_detect_match}
\title{Detect the presence of a pattern in a string}
\usage{
str_detect_match(strings, pattern, fixed = FALSE)
}
\arguments{
\item{strings}{A character vector, where each element of the vector is a character string.}

\item{pattern}{A single character string to be searched for in each element of \code{strings}.
By default, \code{pattern} is interpreted as a regular expression (regex). If the \code{fixed} argument is set to \code{TRUE},
\code{pattern} will be treated as a literal string to be matched exactly.}
}
\value{
A logical vector indicating the presence of each pattern in a string.
}
\description{
These functions search for patterns in strings
}
\examples{
str_detect_match(c("apple", "banana", "cherry"), "a")
#> [1]  TRUE  TRUE FALSE
}
