% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string-pattern-matching.R
\name{str_detect_match}
\alias{str_detect_match}
\title{Detect the presence of a pattern in a string}
\usage{
str_detect_match(string, pattern)
}
\arguments{
\item{string}{A character vector of strings to search in.}

\item{pattern}{The pattern to look for.}
}
\value{
A logical vector indicating the presence of each pattern in a string.
}
\description{
These functions search for patterns in strings
}
\examples{
str_detect_match(c("apple", "banana", "cherry"), "a")
#> [1]  TRUE  TRUE FALSE
}
