% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string-transformation.R
\name{str_indent}
\alias{str_indent}
\title{Add characters to the ends of strings}
\usage{
str_indent(strings, times = 1, side = c("left", "right", "both"), indent = " ")
}
\arguments{
\item{strings}{A character vector, where each element of the vector is a character string.}

\item{times}{The number of times the \code{indent} should be repeated, on each side specified.}

\item{side}{Which side of the string should the indent be applied, one of
"left", "right" or "both". The left side is the default.}

\item{indent}{The characters to be used to indent.}
}
\value{
A character vector of the same length as \code{strings}, with each
element indented as specified by the \code{times}, \code{indent}, and \code{side} arguments.
}
\description{
Add characters to the ends of strings
}
\examples{
str_indent(c("Hello", "World"), 3)
#> [1] "   Hello" "   World"
str_indent(c("Hello", "World"), 3, ".")
#> [1] "...Hello" "...World"
str_indent(c("Hello", "World"), 3, "-", "both")
#> [1] "---Hello---" "---World---"

# Being extra
"wow" |>
  str_indent(side = "both") |>
  str_indent(3, indent = "->>") |>
  str_indent(3,  side = "right", indent = "<<-") |>
  str_indent(2, side = "both", indent = "-||-") |>
  str_indent(2, indent = "[") |>
  str_indent(2, side = "right", indent = "]")
#> [1] "[[-||--||-->>->>->> wow <<-<<-<<--||--||-]]"
}
\seealso{
\code{\link[=str_concat]{str_concat()}} and \code{\link[=str_glue]{str_glue()}} for combining strings together.

Base \code{\link[=paste0]{paste0()}} and \code{\link[=strrep]{strrep()}} used by this function.
}
