% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string-pattern-matching.R
\name{str_locate_first}
\alias{str_locate_first}
\alias{str_locate_all}
\title{Locate Pattern Occurrences in Strings}
\usage{
str_locate_first(string, pattern)

str_locate_all(string, pattern)
}
\arguments{
\item{string}{A character vector of strings to search in.}

\item{pattern}{A character string containing a regular expression.}
}
\value{
\code{str_locate_first()} returns a two-column matrix with the start and end positions
of the first match. There is a row for each string. \code{str_locate_all()} returns a list of matrices.
There is a matrix for each string and a row for each match.
If no match is found, NA values are returned.
}
\description{
These functions find occurrences of a pattern in strings.

\code{str_locate_first()} finds the first occurrence of a pattern in a string.
\code{str_locate_all()} finds all occurrences of a pattern in each string of the input vector.
}
\examples{
str_locate_first("Hello world", "world")
#>      start end
#> [1,]     7  11
str_locate_first(c("Hello world", "Goodbye world"), "o")
#>      start end
#> [1,]     5   5
#> [2,]     2   2

str_locate_all(c("Hello world", "Goodbye world"), "world")
#> [[1]]
#>      start end
#> [1,]     7  11
#>
#> [[2]]
#>      start end
#> [1,]     9  13
str_locate_all(c("Hello world", "Goodbye world"), "o")
#> [[1]]
#> start end
#> [1,]     5   5
#> [2,]     8   8
#>
#> [[2]]
#> start end
#> [1,]     2   2
#> [2,]     3   3
#> [3,]    10  10

}
