% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string-pattern-matching.R
\name{str_remove_first}
\alias{str_remove_first}
\alias{str_remove_all}
\title{Remove Patterns from Strings}
\usage{
str_remove_first(strings, pattern, fixed = FALSE)

str_remove_all(strings, pattern, fixed = FALSE)
}
\arguments{
\item{strings}{A character vector. Each element of this vector is a string that the function will process.}

\item{pattern}{A single character string containing a regular expression (regex) pattern to match against the elements in \code{strings}.
The regex used in \code{{suitestrings}} are Perl-like, for the extended features and syntax.
This is achieved by setting \verb{perl = TRUE, fixed = fixed} in the underlying base R pattern matching functions.
See R's \code{\link{regexp}} documentation for more details.}
}
\value{
A character vector of the same length as \code{strings}, with the specified pattern removed.
For \code{str_remove_first()}, only the first occurrence of the pattern in each string is removed.
For \code{str_remove_all()}, all occurrences of the pattern in each string are removed.
}
\description{
\code{str_remove_first()} removes the first occurrence of a pattern in each string.
\code{str_remove_all()} removes all occurrences of a pattern in each string.
}
\examples{
str_remove_first("Hello world", "o")
#> [1] "Hell world"

str_remove_all("Hello world", "o")
#> [1] "Hell wrld"

}
\seealso{
\code{\link{str_replace_first}}, \code{\link{str_replace_all}} for replacing patterns with specific text.
}
