% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string-pattern-matching.R
\name{str_replace_first}
\alias{str_replace_first}
\alias{str_replace_all}
\title{Replace parts of a string with new text.}
\usage{
str_replace_first(string, pattern, replacement)

str_replace_all(string, pattern, replacement)
}
\arguments{
\item{string}{A character vector of strings to edit.}

\item{pattern}{A single string containing a regular expression pattern to match.}

\item{replacement}{The replacement string}
}
\value{
\code{str_replace()} Returns an altered character vector of equal length to \code{string},
with the first match in each string replaced.

\code{str_replace_all()} Returns an altered character vector of equal length to \code{string},
with every match in each string replaced. Regular expression matches are non-overlapping.
}
\description{
Replace parts of a string with new text.
}
\examples{
str_replace_first("Hello world!", "o", "ooo")
#> [1] "Hellooo world!"

str_replace_all("Hello world!", "o", "ooo")
#> [1] "Hellooo wooorld!"

}
