% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string-pattern-matching.R,
%   R/vector-pattern-matching.R
\name{str_split_all}
\alias{str_split_all}
\alias{chr_split_all}
\title{Split strings by a pattern}
\usage{
str_split_all(string, pattern)

chr_split_all(string, pattern)
}
\arguments{
\item{string}{A character vector of strings to be split.}

\item{pattern}{A character string containing a regular expression (or a fixed string) to split the input \code{string} by.}
}
\value{
For \code{str_split}: A list of the same length as \code{string}, with each element being a vector of substrings obtained by splitting the corresponding element of \code{string}.

For \code{chr_split}: A single character vector containing all substrings obtained by splitting each element of \code{string} and concatenating them.
}
\description{
\code{str_split} splits each string in the input vector into substrings based on a pattern.
\code{chr_split} splits all strings in the input vector and concatenates the results into a single character vector.
}
\examples{
str_split(c("one,two,three", "abc,def,ghi"), ",")
#> [[1]]
#> [1] "one"   "two"   "three"
#>
#> [[2]]
#> [1] "abc" "def" "ghi"

chr_split(c("one,two,three", "abc,def,ghi"), ",")
#> [1] "one"   "two"   "three" "abc"   "def"   "ghi"

}
\seealso{
\code{\link{strsplit}} for the base R function.
}
